+++ Admin/slider/models.py
@@ -8,5 +8,6 @@ class GalleryModel(models.Model):
 
     def delete(self, *args, **kwargs):
         # Delete the image file when the product is deleted
-        os.remove(self.slider_img.path)
+        if self.slider_img and os.path.exists(self.slider_img.path):
+            os.remove(self.slider_img.path)
         super(GalleryModel, self).delete(*args, **kwargs)
